V0.9.1
Updated PlatformDefs to include typedefs so it is safer with templates.
Updated PluginLoader to also use those same typedefs.
Updated PluginRegistry to use map.emplace and check for duplicate key.
It now throws an exception on a duplicate key to help provide a way for
the end user to easily handle plugin modules that do not conform to the
framework requirements.
Finally, alphabetized the includes in PluginRegistry.h (yay OCD!)